Update cpp-widget.php
This change introduces a second option to this widget that shortens the length of the displayed information. When selected, the widget shortens the sign to the standard 3-characters, shortens the planet names down to 3 or 4 characters and drops the precision of the second.  ** Note that I did not change any other files, thus the internationalization of the shortened names may be broken. Would love to learn how to fix that.